build: core: Disable Jack & Jill
Jack is deprecated as of Android O, and only a few legacy packages
actually depend on it. These packages can in turn be compiled using
modern methods without any modifications whatsoever.

There is a lot more to this change... It marks something that hasn't
been as heavy since when we switched TCs back in the LP era.

Change-Id: I99ebe4ddad3d5518bebfbecca4183a150519de27